requirements.txt
alembic==0.9.3
amqp==1.4.9
angular-gettext-babel==0.3
aniso8601==1.1.0
anyjson==0.3.3
arrow==0.13.2
asciitree==0.3.3
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.6.0
autosemver==0.5.5
Babel==2.5.3
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.lzma==0.0.6
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
binaryornot==0.4.0
bleach==3.2.1
blinker==1.4
boto3==1.17.102
botocore==1.20.102
celery==3.1.26.post2
Cerberus==0.7
certifi==2021.10.8
cffi==1.11.5
chardet==3.0.4
click==6.7
configobj==5.0.6
constantly==15.1.0
contextlib2==0.5.4
cookiecutter==1.4.0
cryptography==2.3
cssselect==1.0.1
decorator==4.0.10
dojson==1.4.0
dulwich==0.19.15
ecdsa==0.13
elasticsearch==7.10.1
elasticsearch-dsl==7.4.0
enum34==1.1.6
Fabric==1.12.0
fixture==1.5.8
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Alembic==2.0.1
Flask-Assets==0.12
Flask-Babel==0.11.1
Flask-BabelEx==0.9.3
Flask-Breadcrumbs==0.4.0
Flask-Cache==0.13.1
Flask-Caching==1.7.2
Flask-CeleryExt==0.3.1
Flask-CLI==0.4.0
Flask-Collect==1.2.2
Flask-Cors==3.0.3
Flask-Email==1.4.4
Flask-Gravatar==0.4.2
Flask-IIIF==0.2.0
Flask-KVSession==0.6.2
Flask-Limiter==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Menu==0.7.0
Flask-OAuthlib==0.9.5
Flask-Principal==0.4.0
Flask-Registry==0.2.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-Security==3.0.0
Flask-SQLAlchemy==2.2
flask-webpackext==1.0.0
Flask-WTF==0.14.2
flower==0.9.2
fs==0.5.4
ftfy==4.4.3
ftputil==3.3.1
funcsigs==1.0.2
functools32==3.2.3.post2
furl==0.5.6
future==0.16.0
futures==3.2.0
gunicorn==19.6.0
html5lib==1.0.1
hyperlink==17.3.0
idna==2.6
IDUtils==1.1.1
incremental==17.5.0
infinity==1.4
inspire-crawler==1.1.2
inspire-dojson==61.1.11
inspire-schemas==61.4.1
inspire-utils==3.0.4
intbitset==2.3.0
intervals==0.8.0
invenio-access==1.0.2
invenio-accounts==1.0.2
invenio-admin==1.1.1
invenio-assets==1.1.2
invenio-base==1.0.2
invenio-cache==1.0.0
invenio-celery==1.0.0b2
invenio-config==1.0.1
invenio-db==1.0.3
invenio-files-rest==1.0.0b1
invenio-i18n==1.0.0
invenio-indexer==1.1.0
invenio-jsonschemas==1.0.1
invenio-logging==1.1.1
invenio-mail==1.0.2
invenio-oaiharvester==1.0.0a4
invenio-oaiserver==1.1.0
invenio-oauth2server==1.0.3
invenio-oauthclient==1.1.2
invenio-pidstore==1.0.0
invenio-query-parser==0.6.0
invenio-records==1.0.1
invenio-records-files==1.0.0a11
invenio-records-rest==1.5.0
invenio-records-ui==1.0.0b2
invenio-rest==1.1.3
invenio-search[elasticsearch7]==1.2.0
invenio-search-ui==1.0.1
invenio-theme==1.0.0b2
invenio-utils==0.2.0
invenio-workflows==7.0.6
invenio-workflows-files==1.0.0
invenio-workflows-ui==2.0.19
ipaddr==2.2.0
ipaddress==1.0.19
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
isbnid==0.4.4.1
isbnid-fork==0.5.2
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
jinja2-time==0.2.0
jsmin==2.2.2
jsonpatch==1.16
jsonpointer==1.10
jsonref==0.1
jsonresolver==0.2.1
jsonschema==2.6.0
kombu==3.0.37
langdetect==1.0.7
limits==1.2.0
LinkHeader==0.4.3
lxml==4.2.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.2
maxminddb==1.3.0
maxminddb-geolite2==2018.308
mock==2.0.0
more-itertools==4.3.0
msgpack-python==0.5.6
nameparser==0.5.8
node-semver==0.1.1
oauthlib==2.1.0
ordereddict==1.1
orderedmultidict==0.7.10
paramiko==1.18.5
parsel==1.2.0
passlib==1.7.1
pathlib2==2.3.0
pbr==1.10.0
pdfminer==20140328
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.1
pluggy==0.12.0
poyo==0.4.0
prompt-toolkit==1.0.8
psycopg2==2.7
psycopg2-binary==2.8.3
ptyprocess==0.5.1
py==1.7.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycountry==17.9.23
pycparser==2.18
pycrypto==2.6.1
PyDispatcher==2.0.5
Pygments==2.1.3
PyJWT==1.6.1
pynpm==0.1.1
pyOpenSSL==17.5.0
pyPEG2==2.15.2
pytest-runner==2.11.1
python-dateutil==2.6.1
python-editor==1.0.3
python-scrapyd-api==2.0.1
pytz==2018.4
pywebpack==1.0.0
PyYAML==3.13
queuelib==1.4.2
raven==5.32.0
redis==2.10.6
regex==2018.8.29
requests==2.20.0
requests-oauthlib==1.0.0
rfc3987==1.3.8
scandir==1.5
Scrapy==1.4.0
scrapyd==1.2.0
scrapyd-client==1.0.1
service-identity==17.0.0
Sickle==0.6.3
simplegeneric==0.8.1
simplejson==3.8.2
simplekv==0.11.8
singledispatch==3.4.0.3
six==1.12.0
speaklater==1.3
SQLAlchemy==1.2.6
SQLAlchemy-Continuum==1.3.1
sqlalchemy-mptt==0.2.1
SQLAlchemy-Utils==0.33.2
tornado==5.1.1
traitlets==4.3.1
 --install-option="--no-deps" Twisted==17.5.0
ua-parser==0.8.0
Unidecode==1.0.23
uritools==2.2.0
urllib3==1.24.3
validators==0.11.0
vine==1.1.4
w3lib==1.17.0
wcwidth==0.1.7
webargs==5.5.0
webassets==0.12.1
webencodings==0.5.1
Werkzeug==0.14.1
whichcraft==0.4.0
workflow==2.1.3
WTForms==2.1
WTForms-Alchemy==0.16.1
WTForms-Components==0.10.0
zope.interface==4.4.2
flask-shell-ipython==0.4.1
